Refactoring: Split secure.*pp into common, ledger and blockstore, and move working_path to utility, where it was declared anyway (#662)
