+++ rai/core_test/uint256_union.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 
+#include <rai/common.hpp>
 #include <rai/lib/interface.h>
-#include <rai/secure.hpp>
 
 #include <ed25519-donna/ed25519.h>
