+++ rai/node/node.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <rai/ledger.hpp>
 #include <rai/lib/work.hpp>
 #include <rai/node/bootstrap.hpp>
 #include <rai/node/wallet.hpp>
