+++ rai/node/utility.cpp
@@ -1,10 +1,28 @@
-#include <rai/node/utility.hpp>
-
 #include <rai/lib/interface.h>
+#include <rai/node/utility.hpp>
+#include <rai/node/working.hpp>
 
 #include <lmdb/libraries/liblmdb/lmdb.h>
 
 #include <ed25519-donna/ed25519.h>
+
+boost::filesystem::path rai::working_path ()
+{
+	auto result (rai::app_path ());
+	switch (rai::rai_network)
+	{
+		case rai::rai_networks::rai_test_network:
+			result /= "RaiBlocksTest";
+			break;
+		case rai::rai_networks::rai_beta_network:
+			result /= "RaiBlocksBeta";
+			break;
+		case rai::rai_networks::rai_live_network:
+			result /= "RaiBlocks";
+			break;
+	}
+	return result;
+}
 
 boost::filesystem::path rai::unique_path ()
 {
