+++ rai/node/working.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <rai/secure.hpp>
+#include <rai/common.hpp>
 
 namespace rai
 {
